requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
imageio==2.9.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.10.1
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
